fix: corrected docs and a few tests
