+++ src/facades/likes.js
@@ -21,9 +21,9 @@ export default {
    * });
    *
    * @param {Object} options
-   * @param {!String} options.userId - A user ID.
+   * @param {?String} options.userId - A user ID.
    * @param {!String} options.apiKey - A store api key.
-   * @param {!String} options.secretKey - A store secret key.
+   * @param {?String} options.secretKey - A store secret key.
    * @param {!String} options.itemType - The type of item, can be 'product' or 'collection'.
    * @param {?String} options.productFormat - The format in which the lists products
    * should be returned.
@@ -80,9 +80,9 @@ export default {
    * });
    *
    * @param {Object} options
-   * @param {!String} options.userId - A user ID.
+   * @param {?String} options.userId - A user ID.
    * @param {!String} options.apiKey - A store api key.
-   * @param {!String} options.secretKey - A store secret key.
+   * @param {?String} options.secretKey - A store secret key.
    * @param {!String} options.itemType - The type of item, can be 'product' or 'collection'.
    * @param {?Boolean} options.showOnlyAvailable - Whetever unavailable items should be
    * used toward the total.
@@ -131,9 +131,9 @@ export default {
    *
    * @param {Object} options
    * @param {!String} options.itemId - An item ID.
-   * @param {!String} options.userId - A user ID.
+   * @param {?String} options.userId - A user ID.
    * @param {!String} options.apiKey - A store api key.
-   * @param {!String} options.secretKey - A store secret key.
+   * @param {?String} options.secretKey - A store secret key.
    * @param {!String} options.itemType - The type of item, can be 'product' or 'collection'.
    * @param {?String} options.productFormat - The format in which the lists products
    * should be returned.
@@ -186,9 +186,9 @@ export default {
    *
    * @param {Object} options
    * @param {!String} options.itemId - An item ID.
-   * @param {!String} options.userId - A user ID.
+   * @param {?String} options.userId - A user ID.
    * @param {!String} options.apiKey - A store api key.
-   * @param {!String} options.secretKey - A store secret key.
+   * @param {?String} options.secretKey - A store secret key.
    * @param {!String} options.itemType - The type of item, can be 'product' or 'collection'.
    * @returns {Promise} - A message of success
    */
