+++ test/facades/likes.spec.js
@@ -91,7 +91,7 @@ describe('LikesFacade', function() {
       expect(response).to.eventually.deep.equal(expectedResponse);
     });
 
-    it('should return null when the server returns an error', function() {
+    it('should reject when the server returns an error', function() {
       const options = {
         userId: 'user123456',
         apiKey: 'apiKeyFake',
@@ -108,64 +108,6 @@ describe('LikesFacade', function() {
       server.respondImmediately = true;
 
       const response = LikesFacade.getCount(options);
-      expect(response).to.be.rejected;
-    });
-  });
-
-  describe('getFrequency', function() {
-    beforeEach(function() {
-      server = sinon.createFakeServer();
-    });
-
-    afterEach(function() {
-      server.restore();
-    });
-
-    it('should return an array of item frequencies', function() {
-      const options = {
-        itemIds: ['123', '456', '789'],
-        userId: 'user123456',
-        apiKey: 'apiKeyFake',
-        secretKey: 'secretKeyFake',
-        itemType: 'product',
-      };
-
-      const expectedResponse = [
-        { id: 123, frequency: 98 },
-        { id: 456, frequency: 76 },
-        { id: 789, frequency: 54 },
-      ];
-
-      server.respondWith(
-        'GET',
-        /\/engage\/wishlist\/v3\/likes\/items\/frequency/,
-        [200, { 'Content-Type': 'application/json' }, JSON.stringify(expectedResponse)],
-      );
-
-      server.respondImmediately = true;
-
-      const response = LikesFacade.getFrequency(options);
-      expect(response).to.eventually.deep.equal(expectedResponse);
-    });
-
-    it('should reject when the server returns an error', function() {
-      const options = {
-        itemIds: '[\'123\', \'456\', \'789\']',
-        userId: 'user123456',
-        apiKey: 'apiKeyFake',
-        secretKey: 'secretKeyFake',
-        itemType: 'product',
-      };
-
-      server.respondWith(
-        'GET',
-        /\/engage\/wishlist\/v3\/likes\/items\/frequency/,
-        [500, { 'Content-Type': 'application/json' }, 'Internal server error'],
-      );
-
-      server.respondImmediately = true;
-
-      const response = LikesFacade.getFrequency(options);
       expect(response).to.be.rejected;
     });
   });
@@ -257,7 +199,7 @@ describe('LikesFacade', function() {
       expect(response).to.eventually.deep.equal(expectedResponse);
     });
 
-    it('should return an empty object when the server returns an error', function() {
+    it('should reject when the server returns an error', function() {
       const options = {
         userId: 'user123456',
         apiKey: 'apiKeyFake',
